+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Yandex support.
+- Speech synthesis settings.
 - New convenient languages menu.
 - An option for CLI to read source text from file.
 - An option for CLI to read stdin.